Allow setting download_byte_limit
It can be useful when working with downloads
from untrusted sources in combination with
calls like `get_file`. If we can limit the
amount of data received we can prevent servers
suddenly sending us tons and tons of data whereas
in fact we only expect a few KB. By default
the option is disabled.